Update price and weight types in product entities to float64
Changed `Price`, `Weight`, and related fields in product entities to `float64` for improved precision. Updated all affected entities and validations accordingly. Simplified imports in handlers.

+++ entity/product-data.go
@@ -8,11 +8,11 @@ import (
 type ProductData struct {
 	Uid           string   `json:"product_uid" validate:"required"`
 	Article       string   `json:"article"`
-	Price         float32  `json:"price"`
+	Price         float64  `json:"price"`
 	Quantity      int      `json:"quantity"`
 	Manufacturer  string   `json:"manufacturer"`
 	Active        bool     `json:"active"`
-	Weight        float32  `json:"weight"`
+	Weight        float64  `json:"weight"`
 	WeightClassId int      `json:"weight_class_id"`
 	Categories    []string `json:"categories"`
 	Attributes    []string `json:"attributes"`
